+++ example/lib/computed.g.dart
@@ -13,7 +13,7 @@ const __myAgeMacroHash = 116800019;
 const myAge = 'MyAge: 32';
 
 const __randomNumberMacroHash = 340677358;
-const randomNumber = 782;
+const randomNumber = 484;
 
 const __rawCodeMacroHash = 285619463;
 final rawCode = () {