+++ lib/src/analyzer/utils/hash.dart
@@ -1,5 +1,11 @@
+import 'dart:io';
+
 import 'package:hashlib/hashlib.dart';
 
 int generateHash(String value) {
   return xxh32code(value);
+}
+
+int hashFile(File file) {
+  return xxh3.fileSync(file).number();
 }