+++ lib/src/common/common.dart
@@ -10,11 +10,17 @@ enum ConnectionStatus { connecting, connected, disconnected }
 /// The dart binary from the current sdk.
 final dartBinary = p.join(sdkBin, 'dart');
 
+/// The flutter binary from the current sdk.
+final flutterBinary = p.join(flutterPath, 'flutter');
+
 /// The path to the sdk bin directory on the current platform.
 final sdkBin = p.join(sdkPath, 'bin');
 
 /// The path to the sdk on the current platform.
 final sdkPath = p.dirname(p.dirname(Platform.resolvedExecutable));
+
+/// The path to the flutter cli on the current platform.
+final flutterPath = p.dirname(p.dirname(sdkPath));
 
 String get homeDir {
   if (Platform.isMacOS) {